Fix errors in StandardAnchorTags tool (#255)
* Fix TOC header line, link text

The "C#" should not appear in an anchor.

* Fix for headers that have multiple spaces

I'm sure we'll run into this scenario again.

First change: Add the "Remove Empty Entries" option to string split options. This means any header lines with repeated spaces are ignored.

Second change: Return null if the line is blank, or whitespace. That's obviously not a header.

* remove test code to break in debugger.